refactor: add garden ar view model
